[#30] refactor: types 폴더 이동에 따른 import 수정
